refatoração finalizada. E trocados ids de Integer para Long

+++ src/main/java/my_app/core/db/dto/CompraDto.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record CompraDto(
         String produtoCod,
         BigDecimal precoCompra,
-        Integer fornecedorId,
+        Long fornecedorId,
         BigDecimal quantidade,
         BigDecimal descontoEmReais,
         String tipoPagamento,
